topic related admin operations refactor
